Fix: Wrong values were written on multi-bit sets
We had a logic error that assumed that we can just use the OR operator
when setting multiple bits at once just like it is done in SET_BIT.
This would lead to wrong values when trying to overriding a set bit with
a zero.

Example:
bitfield = 0b1001, value = 0b01, bit = 2.
`bitfield |= (value << bit)` would result in 0b1101 instead of the
expected 0b0101.

To fix this we first clear all bits that will be touched. For this we
naturally need the length of the value now.