scheduler_cycle on local timer irq
The handler now calls the scheduler_cycle function, which, in theory
means, that the scheduler is now interrupt based. Unfortunately, as we
don't have any context switching right now, this means, that the threads
callback function is now being called with heightened privileges and the
I-bit in the CPSR set. This means that all further IRQs are now masked
and won't actually throw the interrupt.