Infinite loop in threads callback function
The simple threads are now running infinitely. This ensures that we can
see the switching and saving of contexts as we can then see, that
whenever the thread is running, it is counting up a number starting
where it left of the last time the same thread was running.

As we are not doing context switching and saving yet, this means that
the first thread to be called by the scheduler just runs forever. Also
because the I-bit mask in the CPSR is set, it cannot be interrupted by
another IRQ (i.e. the local timer) and thus really just runs forever.