Fix bug in scheduler_cycle when queue is full
Change order of queueing and dequeuing so that we do not try to overfill
the queue even though we would remove a thread in a second.

Remove TODO of volatile pointer for running_thread. We kind of remove
it from the queue, so a pointer does not make much sense.