Change waiting_queue to be index based
Restructure the queue to be index based. by that we move a lot less
memory around on every context switch and it helps on fixing the bug
that currently we do not have a way of reassigning the stacks to a new
thread after a thread was killed/terminated that used that stack.