Fixes and error handling in scheduler
Declare functions that are not in the header as static.
Use return values like `-1` and `0` and log error instead of crashing
the full system.
More error handling in edge cases.
Remove outdated comments.

Push back index to `free_indices_q` if we cannot push the new thread in
`schedule_thread`.

Fix that uninitialized threads are pushed pack and not lost. Continue to
pop until we have an initialized thread or we went through the full
queue.

Do not write the registers of the null-thread.
Set `lr` of null-thread to endless loop on init.