WIP Design of the interface for the mmu
L1 table is placed by the linker. This might still be faulty but does
not throw an error yet.
Stub functions for the interface and a finished init_mmu function which
calls these stub functions.
As these stub functions still always assert (as they are unfinished) I
removed the call to the function in start_kernel. The call would work
though.