Rename mem* functions in user mode
We do not want memcpy to be declared as weak and then overwritten by the
implementation in user mode as the kernel then tries to execute user
code.
Rename the user functions for now.
TODO: Find a solution (linker script?) that makes having two memcpys in
kernel and user space possible.