Refactor index_queue
Index queues need to be different depending on the size. As we can not
yet allocate memory, we can't have a unified queue data type, so for
now, we have a macro which generates the code necessary for an
index_queue of a given size. Unfortunately this means that these can
never be used in header files, or anything with a "global" scope (global
meaning, it can be seen from other c- or asm-source-files) as these
would conflict with the local implementations as the names are always
the same.

This is not ideal and will be fixed later on in the development process,
although we have not yet decided how.