FIX: create_thread
When creating a new process the l1 entry still points to the l2 table of
the currently running thread. This means we do not have access to the
stack of the thread in creation. That means however, that we cannot copy
over the memory from the old stack to the new stack.

To fix this we need to save the arguments in the kernel space, change
the l2_table to the one of the process of the new thread, whose memory
must also be initialized beforehand, then copy it into the stack of the
new thread and then switch the memory back to allow access of the stack
of the calling thread.