Change switch_memory to also change udata memory
As all the processes are supposed to have their own udata segment we
need to switch these out to point to their respective physical memory.

IMPORTANT: For right now, this only works with small udata segments and
a little used stack as the memory gets copied into the same MB in
memory, where the stacks are also located.