Make index_queue independent of its buffer size
Instead of declaring at compile time how big the buffer should be (since
we save it redundantly in size anyway) the index_queue struct that just
keeps a pointer now. Because of that the buffer has to be declared
separately and passed to `init()` but that is nothing that a small macro
wrapper could not fix.