Put udata segment into same virtual mem as ustacks
The user data and bss segments now share the same virtual memory as the
user stacks. As the user and data segments are now inside an l2 table
this allows for a more fine grained control over the virtual memory
addresses and as a by product cleaned up quite some code involving the
memory as there is now only one L1 entry to manage per process.