WIP Fix delay on thread creation + empty scheduler
Currently when creating a new thread and the null_thread is running,
there is a short delay until the timer interrupt fires and the new
thread runs.
This should fix this, although it does not seem to do it yet.
It is also currently missing the possibility to fix the delay on boot as
we do not have a registers struct.